Refactor: appendTodo() 함수 분리
